[Gold V] Title: 최소비용 구하기, Time: 116 ms, Memory: 4628 KB -BaekjoonHub

+++ 백준/Gold/1916. 최소비용 구하기/최소비용 구하기.cc
@@ -31,7 +31,7 @@ int main() {
   
   while(!pq.empty()) {
     cost = -pq.top().first;
-    long long current = pq.top().second;
+    int current = pq.top().second;
     pq.pop();
     
     if (dist[current] < cost) continue;
